Created history in database and a tabbed view in main window.
Functions for handling files in history is added and exisiting
functions are changed to be able to handle history.
Move projects from/to history is added.

Main window: Delete project is added under file menu.
A tabbed window is added for view of both active projects
and project history.

About window is fixed and GPLV3 file is added to be able to
be read in when about dialog is shown.

 On branch master
 Changes to be committed:
   (use "git reset HEAD <file>..." to unstage)

	modified:   files.cpp
	modified:   files.h
	new file:   haprdb/history/987654321.project
	new file:   haprdb/history/98765432101_987654321.activity
	new file:   haprdb/history/98765432102_987654321.activity
	new file:   haprdb/history/98765432103_987654321.activity
	new file:   images/GPLV3
	new file:   images/HandleProject_300x70.gif
	modified:   main.cpp
	modified:   makefile
	new file:   ui/about.cpp
	new file:   ui/about.h
	modified:   ui/mainwindow.cpp
	modified:   ui/mainwindow.h