Added a NewEditDialog, a SplashWindow and a ProjectWindow
NewEditDialog is to create new/edit project information,
from MainWindow and from ProjectWindow. Still missing
update in main window when a new project is created.
SplashWindow is coded but is not in use and is not called
from code.
ProjectWindow is still just a base window with a title,
is not completed.

 On branch master
 Changes to be committed:
   (use "git reset HEAD <file>..." to unstage)

	modified:   main.cpp
	modified:   makefile
	modified:   ui/mainwindow.cpp
	modified:   ui/mainwindow.h
	new file:   ui/neweditdialog.cpp
	new file:   ui/neweditdialog.h
	new file:   ui/projectwindow.cpp
	new file:   ui/projectwindow.h
	new file:   ui/splashwindow.cpp
	new file:   ui/splashwindow.h