Import functionality added for main window.
Added function for import in files.
Project that is imprted must have been exported from
Handle Project. All functionality is based on that the
tar file is built up in the correct way.
tar and gzip is used to unpack project.
If project id already exists in database a new is
generated from project class, and this is added in project.
Main list with projects in main window is updated.

 On branch master
 Changes to be committed:
   (use "git reset HEAD <file>..." to unstage)

	modified:   files.cpp
	modified:   files.h
	modified:   project/project.cpp
	modified:   project/project.h
	modified:   ui/mainwindow.cpp